cleanup of extra logging setup
