+++ readahead.py
@@ -1,23 +1,12 @@
 #!/usr/bin/env python
 from __future__ import absolute_import
 import os
-import logging
-import sys
 
 import argparse
 import redis
 import yaml
 
 from tscached.shadow import perform_readahead
-
-
-def setup_logging():
-    logger = logging.getLogger()
-    handler = logging.StreamHandler()
-    formatter = logging.Formatter("%(asctime)s [%(threadName)-12.12s] [%(levelname)-5.5s]  %(message)s")
-    handler.setFormatter(formatter)
-    logger.addHandler(handler)
-    logger.setLevel(logging.DEBUG)
 
 
 if __name__ == '__main__':
